refactor(api): 나이스 상태 코드 처리 함수 분리
